pyproject.toml
[project]
name = "flit-core"
version = "0.1.0"
description = "Flit PKM backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi >=0.115.0",
    "uvicorn[standard] >=0.30.0",
    "sqlalchemy[asyncio] >=2.0.35",
    "asyncpg >=0.30.0",
    "alembic >=1.13.0",
    "pydantic[email] >=2.9.0",
    "pydantic-settings >=2.5.0",
    "python-dotenv >=1.0.0",
    "python-jose[cryptography] >=3.3.0",
    "python-multipart >=0.0.9",
    "passlib[bcrypt] >=1.7.0",
    "pgvector >=0.4.0",
    "pytest >=8.0.0",
    "pytest-asyncio >=0.23.0",
    "pytest-cov >=4.1.0",
    "httpx >=0.27.0",
    "aiosqlite >=0.20.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
